pyproject.toml
[project]
name = "repl-nix-lioncompanion"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.58.1",
    "pillow>=11.0.0",
    "psutil>=6.1.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "trafilatura>=2.0.0",
    "twilio>=9.4.1",
]
